Move Spark expressions to separate package and improve documentation
