+++ spark-cypher-testing/src/test/scala/org/opencypher/spark/api/io/fs/FSGraphSourceTest.scala
@@ -36,7 +36,7 @@ import org.opencypher.spark.api.io.FileFormat
 import org.opencypher.spark.api.io.util.HiveTableName
 import org.opencypher.spark.api.value.CAPSNode
 import org.opencypher.spark.impl.acceptance.ScanGraphInit
-import org.opencypher.spark.impl.encoders.LongEncoder._
+import org.opencypher.spark.api.value.CAPSEntity._
 import org.opencypher.spark.testing.CAPSTestSuite
 
 class FSGraphSourceTest extends CAPSTestSuite with ScanGraphInit {
