+++ spark-cypher/src/main/scala/org/opencypher/spark/impl/expressions/AddPrefix.scala
@@ -1,0 +1,73 @@
+/**
+ * Copyright (c) 2016-2019 "Neo4j Sweden, AB" [https://neo4j.com]
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Attribution Notice under the terms of the Apache License 2.0
+ *
+ * This work was created by the collective efforts of the openCypher community.
+ * Without limiting the terms of Section 6, any Derivative Work that is not
+ * approved by the public consensus process of the openCypher Implementers Group
+ * should not be described as “Cypher” (and Cypher® is a registered trademark of
+ * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
+ * proposals for change that have been documented or implemented should only be
+ * described as "implementation extensions to Cypher" or as "proposed changes to
+ * Cypher that are not yet approved by the openCypher community".
+ */
+package org.opencypher.spark.impl.expressions
+
+import org.apache.spark.sql.Column
+import org.apache.spark.sql.catalyst.expressions.codegen.{CodegenContext, ExprCode}
+import org.apache.spark.sql.catalyst.expressions.{BinaryExpression, ExpectsInputTypes, Expression, NullIntolerant}
+import org.apache.spark.sql.types.{BinaryType, ByteType, DataType}
+
+/**
+  * Spark expression that adds a byte prefix to an array of bytes.
+  *
+  * @param left expression of ByteType that is added as a prefix
+  * @param right expression of BinaryType to which a prefix is added
+  */
+case class AddPrefix(
+  left: Expression, // byte prefix
+  right: Expression // binary array
+) extends BinaryExpression with NullIntolerant with ExpectsInputTypes {
+
+  override val dataType: DataType = BinaryType
+
+  override val inputTypes: Seq[DataType] = Seq(ByteType, BinaryType)
+
+  override protected def nullSafeEval(byte: Any, byteArray: Any): Any = {
+    AddPrefix.addPrefix(byte.asInstanceOf[Byte], byteArray.asInstanceOf[Array[Byte]])
+  }
+
+  override protected def doGenCode(ctx: CodegenContext, ev: ExprCode): ExprCode =
+    defineCodeGen(ctx, ev, (a, p) => s"(byte[])(${AddPrefix.getClass.getName.dropRight(1)}.addPrefix($a, $p))")
+}
+
+object AddPrefix {
+
+  @inline
+  final def addPrefix(p: Byte, a: Array[Byte]): Array[Byte] = {
+    val n = new Array[Byte](a.length + 1)
+    n(0) = p
+    System.arraycopy(a, 0, n, 1, a.length)
+    n
+  }
+
+  implicit class ColumnPrefixOps(val c: Column) extends AnyVal {
+
+    def addPrefix(prefix: Column): Column = new Column(AddPrefix(prefix.expr, c.expr))
+
+  }
+
+}
